+++ app/models/tangara.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, ForeignKey, Integer, String, Boolean
 from sqlalchemy.orm import relationship
 
-from database import Base
+from db.dev import Base
 
 
 class TangaraModel(Base):
