+++ app/models/vereda.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
 
-from database import Base
+from db.dev import Base
 
 
 class VeredaModel(Base):
